Changed the submission table and the submission form to support creating an event from a submission.
In form.html, changed the "submit event" button to read "update event"
if the submission is being modified rather than created. Also added a
Create button to create an event from this submission.

In the Submission class, added the attribute event_created to indicate
if an Event was created from this Submission.